Fix Win-newlines and prefixed tabs
Windows uses \r\n as newline, and since awk is a Unix tool it recognizes
only \n so \r stays there. Further processing interprets them as newline
which is not what is intended.

Awk removes prefixed whitespace if line is printed with `print` without
arguments. That is not what is inteded.